src/main/java/com/quiz/web/command/authentication/ShowRegisterCommand.java
package com.quiz.web.command.authentication;

import com.quiz.web.command.Command;
import com.quiz.web.utils.Pages;
import com.quiz.web.utils.WebPath;
import jakarta.servlet.ServletException;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;

import java.io.IOException;

public class ShowRegisterCommand implements Command {

    @Override
    public WebPath execute(HttpServletRequest request, HttpServletResponse response) throws ServletException,
            IOException {
        return new WebPath(Pages.REGISTER, WebPath.DispatchType.FORWARD);
    }
}